refactor(source): move File infrastructure into source
