+++ src/Source/DiskFile.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\FileSystem;
+namespace Haspadar\Piqule\Source;
 
 use Haspadar\Piqule\PiquleException;
