+++ src/Source/DiskSourceDirectory.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Haspadar\Piqule\Source;
 
 use FilesystemIterator;
-use Haspadar\Piqule\FileSystem\DiskFile;
 use Haspadar\Piqule\PiquleException;
 use RecursiveDirectoryIterator;
 use RecursiveIteratorIterator;
