+++ src/Source/File.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Haspadar\Piqule\FileSystem;
+namespace Haspadar\Piqule\Source;
 
 interface File
 {
